requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.2.0
Flask==3.1.1
Flask-Session==0.8.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgspec==0.19.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
Werkzeug==3.1.3
